feat: add shimmer
