+++ apps/web/src/components/Shared/Shimmer/PublicationShimmer.tsx
@@ -2,7 +2,13 @@ import type { FC } from 'react';
 
 import UserProfileShimmer from './UserProfileShimmer';
 
-const PublicationShimmer: FC = () => {
+interface PublicationShimmerProps {
+  showActions?: boolean;
+}
+
+const PublicationShimmer: FC<PublicationShimmerProps> = ({
+  showActions = true
+}) => {
   return (
     <div className="space-y-4 p-5">
       <div className="flex justify-between">
@@ -14,11 +20,13 @@ const PublicationShimmer: FC = () => {
           <div className="shimmer h-3 w-7/12 rounded-lg" />
           <div className="shimmer h-3 w-1/3 rounded-lg" />
         </div>
-        <div className="flex gap-7 pt-3">
-          <div className="shimmer h-5 w-5 rounded-lg" />
-          <div className="shimmer h-5 w-5 rounded-lg" />
-          <div className="shimmer h-5 w-5 rounded-lg" />
-        </div>
+        {showActions && (
+          <div className="flex gap-7 pt-3">
+            <div className="shimmer h-5 w-5 rounded-lg" />
+            <div className="shimmer h-5 w-5 rounded-lg" />
+            <div className="shimmer h-5 w-5 rounded-lg" />
+          </div>
+        )}
       </div>
     </div>
   );
