+++ apps/web/src/components/Shared/Snapshot/index.tsx
@@ -1,32 +1,18 @@
 import { LENSTER_POLLS_SPACE, ZERO_ADDRESS } from '@lenster/data';
 import generateSnapshotAccount from '@lenster/lib/generateSnapshotAccount';
 import stopEventPropagation from '@lenster/lib/stopEventPropagation';
-import { Card, Spinner } from '@lenster/ui';
+import { Spinner } from '@lenster/ui';
 import getSnapshotProposal from '@lib/getSnapshotProposal';
 import getSnapshotSpace from '@lib/getSnapshotSpace';
 import { useQuery } from '@tanstack/react-query';
 import type { Proposal, Vote } from '@workers/snapshot-relay';
-import type { FC, ReactNode } from 'react';
+import type { FC } from 'react';
 import { useState } from 'react';
 import { useAppStore } from 'src/store/app';
 
+import Wrapper from '../Embed/Wrapper';
 import Choices from './Choices';
 import Header from './Header';
-
-interface WrapperProps {
-  children: ReactNode;
-  dataTestId?: string;
-}
-
-const Wrapper: FC<WrapperProps> = ({ children, dataTestId = '' }) => (
-  <Card
-    className="mt-3 cursor-auto p-5"
-    dataTestId={dataTestId}
-    onClick={stopEventPropagation}
-  >
-    {children}
-  </Card>
-);
 
 interface SnapshotProps {
   proposalId: string;
